refactored and normalized to mean
